reorganise styleguide entries
