+++ packages/styleguide/src/styleguide.js
@@ -1,10 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from '../App';
-import '../polyfills';
+import App from './components/App';
 
-export * from '../components';
-export * from '../utils';
+export * from './components';
+export * from './utils';
 
 export default function({ config = {}, routes = [] }) {
   ReactDOM.render(
